Zero out unticked scores while importing
We used to use `score_to_answerid($q,$score)` to look up the
*only* the answer corresponding to the score. But when things are
mis-imported, it's hard to undo. So now we zero out the unticked
scores too. This multiplies the number of calls to set_zone... by the
maximum score, but so it goes.

TODO: provide a command-line option to zero out old scores if necessary.